requirements.txt
langchain
langchain-community
PyPDF2
streamlit
faiss-cpu
sentence-transformers
langchain-google-genai